+++ bin/bavard.dart
@@ -1,0 +1,5 @@
+import '../lib/src/cli/cli_runner.dart';
+
+void main(List<String> args) {
+  CliRunner().run(args);
+}